backend/app/requirements.txt
Django==3.2.7
django-cors-headers==3.8.0
django-environ==0.6.0
django-extensions==3.1.3
djangorestframework==3.12.4
gunicorn==19.9.0
ipython==7.27.0
Pillow==8.3.2
